Move new classes to new package
