+++ src/main/java/de/yannikpfeiffer/hadrysconverter/optionloading/Options.java
@@ -1,4 +1,4 @@
-package de.yannikpfeiffer.hadrysconverter;
+package de.yannikpfeiffer.hadrysconverter.optionloading;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
