+++ src/main/java/de/yannikpfeiffer/hadrysconverter/optionloading/StringToPathConverter.java
@@ -1,4 +1,4 @@
-package de.yannikpfeiffer.hadrysconverter;
+package de.yannikpfeiffer.hadrysconverter.optionloading;
 
 import com.fasterxml.jackson.databind.util.StdConverter;
